feat(commons/color): add opacity info
Add opacity info to color contrast remediation text

Closes issue #3928
