+++ lib/checks/color/color-contrast-evaluate.js
@@ -8,6 +8,7 @@ import {
 import {
   getBackgroundColor,
   getForegroundColor,
+  getOpacity,
   incompleteData,
   getContrast,
   getOwnBackgroundColor,
@@ -74,6 +75,7 @@ export default function colorContrastEvaluate(node, options, virtualNode) {
   const bgNodes = [];
   const bgColor = getBackgroundColor(node, bgNodes, shadowOutlineEmMax);
   const fgColor = getForegroundColor(node, false, bgColor, options);
+  const opacity = getOpacity(node, nodeStyle);
   // Thin shadows only. Thicker shadows are included in the background instead
   const shadowColors = getTextShadowColors(node, {
     minRatio: 0.001,
@@ -136,6 +138,7 @@ export default function colorContrastEvaluate(node, options, virtualNode) {
   this.data({
     fgColor: fgColor ? fgColor.toHexString() : undefined,
     bgColor: bgColor ? bgColor.toHexString() : undefined,
+    opacity: opacity ? opacity.toFixed(2) : undefined,
     contrastRatio: truncatedResult,
     fontSize: `${((fontSize * 72) / 96).toFixed(1)}pt (${fontSize}px)`,
     fontWeight: bold ? 'bold' : 'normal',
