+++ lib/commons/color/get-foreground-color.js
@@ -3,7 +3,7 @@ import getBackgroundColor from './get-background-color';
 import incompleteData from './incomplete-data';
 import flattenColors from './flatten-colors';
 import getTextShadowColors from './get-text-shadow-colors';
-import { getNodeFromTree } from '../../core/utils';
+import getOpacity from './get-opacity';
 
 /**
  * Returns the flattened foreground color of an element, or null if it can't be determined because
@@ -81,28 +81,4 @@ function getStrokeColor(nodeStyle, { textStrokeEmMin = 0 }) {
 
   const strokeColor = nodeStyle.getPropertyValue('-webkit-text-stroke-color');
   return new Color().parseString(strokeColor);
-}
-
-function getOpacity(node, nodeStyle) {
-  if (!node) {
-    return 1;
-  }
-
-  const vNode = getNodeFromTree(node);
-  if (vNode && vNode._opacity !== undefined && vNode._opacity !== null) {
-    return vNode._opacity;
-  }
-
-  nodeStyle ??= window.getComputedStyle(node);
-  const opacity = nodeStyle.getPropertyValue('opacity');
-  const finalOpacity = opacity * getOpacity(node.parentElement);
-
-  // cache the results of the getOpacity check on the parent tree
-  // so we don't have to look at the parent tree again for all its
-  // descendants
-  if (vNode) {
-    vNode._opacity = finalOpacity;
-  }
-
-  return finalOpacity;
 }
