+++ lib/commons/color/index.js
@@ -19,3 +19,4 @@ export { default as getRectStack } from './get-rect-stack';
 export { default as hasValidContrastRatio } from './has-valid-contrast-ratio';
 export { default as incompleteData } from './incomplete-data';
 export { default as getTextShadowColors } from './get-text-shadow-colors';
+export { default as getOpacity } from './get-opacity';
